feat: Refactor App.js to hold more state
In this commit, I refactored my App component to hold two
peices of state. One for the userInfo and the other for the user follower.
an axios call was then made to get the information of the user's followers.
The result of the ajax request was then set to state.